restructure pizza component
