+++ src/app/pages/pizza/checkout/checkout.component.html
@@ -1,0 +1,1 @@
+<p>checkout works!</p>
