+++ src/app/pages/pizza/checkout/checkout.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { CheckoutComponent } from './checkout.component';
+import { AntDesignCommonModule } from 'src/app/ant-design-common.module';
+import { ShareModule } from 'src/app/share/share.module';
+import { FormsModule } from '@angular/forms';
+
+@NgModule({
+  declarations: [CheckoutComponent],
+  imports: [
+    AntDesignCommonModule,
+    ShareModule,
+    CommonModule,
+    FormsModule,
+  ]
+})
+export class CheckoutModule { }
