+++ src/app/pages/pizza/create/create.component.css
@@ -1,0 +1,27 @@
+.steps-content {
+    /* margin-top: 16px; */
+    border: 1px dashed #e9e9e9;
+    border-radius: 6px;
+    background-color: #fafafa;
+    min-height: 200px;
+    /* text-align: center; */
+    padding: 10px;
+  }
+
+  .steps-action {
+    margin-top: 24px;
+  }
+
+  button {
+    margin-right: 8px;
+  }
+
+  .center-text {
+    text-align: center;
+  }
+
+  .select-pizza-size {
+    border-radius: 10px;
+    border: 1px solid green;
+    margin-bottom: 5px;
+}
