+++ src/app/store/app.reducer.ts
@@ -1,7 +1,7 @@
 import * as fromWelcomeTask from '../pages/welcome/store/welcome.reducer';
 import * as fromSleepTask from '../pages/sleep/store/sleep.reducer';
 import * as fromToppingReducer from '../pages/pizza/topping/store/topping.reducer';
-import * as fromPizzaReducer from '../pages/pizza/store/pizza.reducer';
+import * as fromPizzaReducer from '../pages/pizza/create/store/pizza.reducer';
 import { ActionReducerMap } from '@ngrx/store';
 
 export interface AppState {
